OneDrive/Desktop/e commerce website/backend/requirements.txt
absl-py==2.3.1
alembic==1.16.4
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.19.1
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.74.0
h5py==3.14.0
idna==3.10
imbalanced-learn==0.13.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.1
keras==3.11.1
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.1.3
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
optuna==4.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==4.25.8
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
shap==0.48.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
sounddevice==0.5.2
SQLAlchemy==2.0.41
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.2
fastapi
uvicorn[standard]
pydantic
pandas
numpy
torch
scikit-learn
python-multipart
